refactor(data): 👷 work - react split into files
- extra refactor I have seen can be done when updating the images
