+++ app/work-experience/komercni-banka/page.tsx
@@ -4,7 +4,7 @@ import { ProjectPageLayoutWrapper } from '@/components/layout/project-page/Proje
 import PageNavigation from '@/components/page-navigation/PageNavigation'
 
 import { sections } from '@/lib/data/projects/work/komercniBanka'
-import { projectsWorkReact } from '@/lib/data/projects/work/projects-overview/workReact'
+import { projectsWorkReact } from '@/lib/data/projects/work/projects-overview/react/workReact'
 
 import { DATA_TEST_IDS } from '@/__tests__/playwright/lib/utils/constants/ids/dataTestIds'
 import { ID } from '@/lib/utils/constants/ids/elementIds'
