+++ lib/data/projects/work/projects-overview/react/workReact.ts
@@ -1,0 +1,8 @@
+import { workReactKomercniBanka } from '@/lib/data/projects/work/projects-overview/react/workReactKomercniBanka'
+import { workReactKooperativa } from '@/lib/data/projects/work/projects-overview/react/workReactKooperativa'
+import { workReactSmartsuppDashboard } from '@/lib/data/projects/work/projects-overview/react/workReactSmartsuppDashboard'
+
+import { Project } from '@/lib/utils/interfaces/interfaces'
+
+export const projectsWorkReact: Project[] = [workReactSmartsuppDashboard, workReactKooperativa, workReactKomercniBanka]
+
